mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.13",
    "flask>=3.1.1",
    "mcp[cli]>=1.9.4",
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
]
